+++ src/test/java/com/wikia/webdriver/common/contentpatterns/MercurySubpages.java
@@ -17,7 +17,7 @@ public class MercurySubpages {
   public static final String QUESTION_MARK = "Question?mark?question";
   public static final String COLON = "Colon:colon:colon";
 
-  public static final String TWITTER_VERBATIM = "TwitterVerbatim";
+  public static final String TWITTER_WIDGET = "TwitterWidget";
 
   /**
    * Legend:
